Added useRoles hook
