+++ pages/[domain]/settings/contributors.tsx
@@ -22,11 +22,7 @@ export default function ContributorSettings () {
   return (
     <>
       <InviteLinkList isAdmin={isAdmin} spaceId={space.id} />
-      {
-        /*
       <RoleAssignment />
-        */
-      }
 
       <ContributorList isAdmin={isAdmin} spaceId={space.id} spaceOwner={space.createdBy} />
     </>
